test: migrate to classic test
